Refactor to help understand tests
